third_party/WebKit/LayoutTests/inspector/elements/styles-3/styles-variables.html

<html>
<head>
<style>
body {
    --a: red;
}

#id1 {
    --b: 44px;
}

#id2 {
    --a: green;
}

#id3 {
    --a: inherit;
}

#id4 {
    --a: var(--z);
}

</style>
<script src="../../../http/tests/inspector/inspector-test.js"></script>
<script src="../../../http/tests/inspector/elements-test.js"></script>
<script>

function test()
{
    InspectorTest.selectNodeAndWaitForStylesWithComputed("id1", step1);
    function step1(node)
    {
        InspectorTest.addResult("==== Computed style for ID1 ====");
        InspectorTest.dumpSelectedElementStyles(false, false);
        InspectorTest.cssModel.computedStylePromise(node.id).then(
            function(style) {
                InspectorTest.addResult("value of --a: " + style.get("--a"));
                InspectorTest.selectNodeAndWaitForStylesWithComputed("id2", step2);
            });
    }

    function step2(node)
    {
        InspectorTest.addResult("==== Computed style for ID2 ====");
        InspectorTest.dumpSelectedElementStyles(false, false);
        InspectorTest.cssModel.computedStylePromise(node.id).then(
            function(style) {
                InspectorTest.addResult("value of --b: " + style.get("--b"));
                InspectorTest.selectNodeAndWaitForStylesWithComputed("id3", step3);
            });
    }

    function step3(node)
    {
        InspectorTest.addResult("==== Computed style for ID3 ====");
        InspectorTest.dumpSelectedElementStyles(false, false);
        InspectorTest.cssModel.computedStylePromise(node.id).then(
            function(style) {
                InspectorTest.addResult("value of --b: " + style.get("--b"));
                InspectorTest.selectNodeAndWaitForStylesWithComputed("id4", step4);
            });
    }

    function step4(node)
    {
        InspectorTest.addResult("==== Computed style for ID4 ====");
        InspectorTest.dumpSelectedElementStyles(false, false);
        InspectorTest.cssModel.computedStylePromise(node.id).then(
            function(style) {
                InspectorTest.addResult("value of --a: " + style.get("--a"));
                InspectorTest.completeTest();
            });
    }
}
</script>
</head>

<body onload="runTest()">
<p>
Tests that computed styles expand and allow tracing to style rules.
</p>

<div id="id1">
<div id="id2">
<div id="id3">
</div>
</div>
</div>
<div id="id4">
</div>

</body>
</html>
